feat: Nelmio + data validation

+++ src/Controller/UserController.php
@@ -3,6 +3,7 @@
 namespace App\Controller;
 
 use App\Entity\User;
+use App\Repository\UserRepository;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\JsonResponse;
@@ -11,15 +12,35 @@ use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\PasswordHasher\Hasher\UserPasswordHasherInterface;
 use Symfony\Component\Routing\Attribute\Route;
 use Symfony\Component\Security\Http\Attribute\IsGranted;
+use Symfony\Component\Validator\Exception\ValidationFailedException;
+use Symfony\Component\Validator\Validator\ValidatorInterface;
+use Symfony\Contracts\Cache\TagAwareCacheInterface;
+use OpenApi\Attributes as OA;
 
 class UserController extends AbstractController
 {
     #[Route('/api/user', name: 'app_user_signup', methods: ['POST'])]
+    #[OA\RequestBody(
+        description: 'Cree un compte utilisateur',
+        required: true,
+        content: new OA\JsonContent(
+            required: ['email', 'password', 'postalCode'],
+            properties: [
+                new OA\Property(property: 'email', type: 'string'),
+                new OA\Property(property: 'password', type: 'string'),
+                new OA\Property(property: 'postalCode', type: 'string'),
+            ],
+            type: 'object'
+        )
+    )]
     public function signup
     (
         Request                $request,
         EntityManagerInterface $em,
-        UserPasswordHasherInterface $passwordHasher
+        UserPasswordHasherInterface $passwordHasher,
+        UserRepository         $userRepository,
+        ValidatorInterface     $validator
+
     ): JsonResponse
     {
         if (!$request->getContent()) {
@@ -27,15 +48,25 @@ class UserController extends AbstractController
         }
         $data = json_decode($request->getContent(), true);
 
-        if (!isset($data['email']) || !isset($data['password']) || !isset($data['postalCode'])) {
-            return new JsonResponse('Invalid request', Response::HTTP_BAD_REQUEST);
+        if ($userRepository->findOneBy(['email' => $data['email']]) ) { //FIXME: Pas ouf niveau securité, à revoir
+            return new JsonResponse('Email déjà utilisé, veuillez en utiliser une autre', Response::HTTP_CONFLICT);
         }
 
         $user = new User();
         $user->setEmail($data['email']);
-        $user->setPassword($passwordHasher->hashPassword($user, $data['password']));
+        if ( $data['password'] !== '' ) {
+            $user->setPassword($passwordHasher->hashPassword($user, $data['password']));
+        } else {
+            $user->setPassword($data['password']);
+        }
         $user->setRoles(['ROLE_USER']);
         $user->setPostalCode($data['postalCode']);
+
+        $errors = $validator->validate($user);
+        if ($errors->count() > 0) {
+            throw new ValidationFailedException($user, $errors);
+        }
+
         $em->persist($user);
         $em->flush();
         return new JsonResponse('Compte utilisateur crée', Response::HTTP_CREATED);
@@ -43,31 +74,69 @@ class UserController extends AbstractController
     }
 
     #[Route('/api/user/{id}', name: 'app_user_delete', methods: ['DELETE'])]
+    #[OA\Parameter(
+        name: 'id',
+        description: 'ID de l\'utilisateur',
+        in: 'path',
+        required: true,
+    )]
+    #[OA\Schema(type: 'integer')]
     #[IsGranted('ROLE_ADMIN')]
     public function delete
     (
-        User                   $user,
-        EntityManagerInterface $em
+        int                   $id,
+        EntityManagerInterface $em,
+        UserRepository         $userRepository
     ): JsonResponse
     {
+        if (!$id) {
+            return new JsonResponse('Veuillez renseigner un ID', Response::HTTP_BAD_REQUEST);
+        }
+
+        $user = $userRepository->find($id);
         if (!$user) {
             return new JsonResponse('Compte utilisateur non trouvé', Response::HTTP_NOT_FOUND);
         }
+
         $em->remove($user);
         $em->flush();
         return new JsonResponse('Compte utilisateur supprimé', Response::HTTP_OK);
     }
 
     #[Route('/api/user/{id}', name: 'app_user_update', methods: ['PUT'])]
+    #[OA\Parameter(
+        name: 'id',
+        description: 'ID de l\'utilisateur',
+        in: 'path',
+        required: true,
+    )]
+    #[OA\Schema(type: 'integer')]
+    #[OA\RequestBody(
+        description: 'Modifier un compte utilisateur',
+        required: true,
+        content: new OA\JsonContent(
+            properties: [
+                new OA\Property(property: 'email', type: 'string'),
+                new OA\Property(property: 'password', type: 'string'),
+                new OA\Property(property: 'postalCode', type: 'string'),
+            ],
+            type: 'object'
+        )
+    )]
     #[IsGranted('ROLE_ADMIN')]
     public function update
     (
-        User                   $user,
+        int                    $id,
         Request                $request,
         EntityManagerInterface $em,
-        UserPasswordHasherInterface $passwordHasher
+        UserPasswordHasherInterface $passwordHasher,
+        TagAwareCacheInterface $cache,
+        UserRepository         $userRepository,
+        ValidatorInterface     $validator
     ): JsonResponse
     {
+        $user = $userRepository->find($id);
+
         if (!$user) {
             return new JsonResponse('Compte utilisateur non trouvé', Response::HTTP_NOT_FOUND);
         }
@@ -76,11 +145,21 @@ class UserController extends AbstractController
             $user->setEmail($data['email']);
         }
         if (isset($data['password'])) {
-            $user->setPassword($passwordHasher->hashPassword($user, $data['password']));
+            if ( $data['password'] !== '' ) {
+                $user->setPassword($passwordHasher->hashPassword($user, $data['password']));
+            } else {
+                $user->setPassword($data['password']);
+            }
         }
         if (isset($data['postalCode'])) {
             $user->setPostalCode($data['postalCode']);
+            $cache->invalidateTags([sprintf("weatherCache%s", $user->getId())]);
         }
+        $errors = $validator->validate($user);
+        if ($errors->count() > 0) {
+            throw new ValidationFailedException($user, $errors);
+        }
+
         $em->flush();
         return new JsonResponse('Compte utilisateur mis à jour', Response::HTTP_OK);
     }
